+++ tests/Integration/Services/AccountManagementServiceTest.php
@@ -11,7 +11,7 @@ use N3XT0R\LaravelWebdavServer\Models\WebDavAccountModel;
 use N3XT0R\LaravelWebdavServer\Exception\Auth\DuplicateUsernameException;
 use N3XT0R\LaravelWebdavServer\Repositories\EloquentAccountRepository;
 use N3XT0R\LaravelWebdavServer\Services\AccountManagementService;
-use N3XT0R\LaravelWebdavServer\Services\AccountUpdateApplier;
+use N3XT0R\LaravelWebdavServer\Services\AccountUpdateService;
 use N3XT0R\LaravelWebdavServer\Tests\DatabaseTestCase;
 
 final class AccountManagementServiceTest extends DatabaseTestCase
@@ -20,7 +20,7 @@ final class AccountManagementServiceTest extends DatabaseTestCase
     {
         $repository = new EloquentAccountRepository($this->app->make(Repository::class));
 
-        return new AccountManagementService($repository, new AccountUpdateApplier($repository));
+        return new AccountManagementService($repository, new AccountUpdateService($repository));
     }
 
     // --- columnMapping ---